refactor(richtexteditor): pass initial value as prop
