+++ src/components/RichTextEditor/index.tsx
@@ -1,12 +1,9 @@
 import React, { useCallback, useContext, useEffect, useMemo } from 'react'
 import isHotkey from 'is-hotkey'
-import { Editable, withReact, Slate } from 'slate-react'
-import {
-  createEditor,
-  Descendant,
-} from 'slate'
+import { Editable, withReact, Slate, ReactEditor } from 'slate-react'
 import { withHistory } from 'slate-history'
 import { BlockButton, MarkButton, toggleMark } from './elements/buttons'
+import { createEditor, BaseEditor, Descendant } from 'slate'
 
 import { Toolbar } from './components'
 import {
@@ -34,19 +31,34 @@ const HOTKEYS = {
   'mod+`': 'code',
 }
 
-const RichTextEditor = () => {
+type CustomElement = { type: 'paragraph' | 'block-quote'; align?: string; children: CustomText[] }
+type CustomText = { text: string; bold?: boolean; italic?: boolean; code?: boolean }
+
+declare module 'slate' {
+  interface CustomTypes {
+    Editor: BaseEditor & ReactEditor
+    Element: CustomElement
+    Text: CustomText
+  }
+}
+
+interface IRichTextEditor {
+  value: Descendant[]
+}
+
+const RichTextEditor = ({value}: IRichTextEditor) => {
   const renderElement = useCallback(props => <Element {...props} />, [])
   const renderLeaf = useCallback(props => <Leaf {...props} />, [])
   const editor = useMemo(() => withHistory(withReact(createEditor())), [])
   const { setSlateValue } = useContext(SlateValueContext)
   useEffect(() => {
-    setSlateValue(JSON.stringify(initialValue))
+    setSlateValue(JSON.stringify(value))
   }, [])
 
   return (
     <Slate
       editor={editor}
-      value={initialValue}
+      value={value}
       onChange={value => {
         const isAstChange = editor.operations.some(
           op => 'set_selection' !== op.type
@@ -161,43 +173,5 @@ const Leaf = ({ attributes, children, leaf }) => {
 
   return <span {...attributes}>{children}</span>
 }
-
-const initialValue: Descendant[] = [
-  {
-    type: 'paragraph',
-    children: [
-      { text: 'This is editable ' },
-      { text: 'rich', bold: true },
-      { text: ' text, ' },
-      { text: 'much', italic: true },
-      { text: ' better than a ' },
-      { text: '<textarea>', code: true },
-      { text: '!' },
-    ],
-  },
-  {
-    type: 'paragraph',
-    children: [
-      {
-        text:
-          "Since it's rich text, you can do things like turn a selection of text ",
-      },
-      { text: 'bold', bold: true },
-      {
-        text:
-          ', or add a semantically rendered block quote in the middle of the page, like this:',
-      },
-    ],
-  },
-  {
-    type: 'block-quote',
-    children: [{ text: 'A wise quote.' }],
-  },
-  {
-    type: 'paragraph',
-    align: 'center',
-    children: [{ text: 'Try it out for yourself!' }],
-  },
-]
 
 export default RichTextEditor
